removed window.location and replace with useNavigate for auth actions, created SavedBooks component
SavedBooks component is operational, styling is decent and everything looks good. Its being used in Profile Component and SavedBooks page. Noticed that if you don't add a size value for grid items they are more responsive. need to fix issue where the state of the books are a bunch of books that have availability set to false